requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dashscope==1.23.8
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.7.0
google==3.0.0
gradio==5.38.0
gradio_client==1.11.0
greenlet==3.2.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.33.4
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-openai==0.3.28
langchain-tavily==0.2.10
langchain-text-splitters==0.3.8
langsmith==0.4.6
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
openai==1.96.1
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
playwright==1.53.0
propcache==0.3.2
protobuf==6.31.1
py==1.11.0
pycparser==2.22
pycryptodome==3.9.9
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-toolbelt==1.0.0
retry==0.9.2
rich==14.0.0
ruff==0.12.3
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.47.1
tenacity==9.1.2
tiktoken==0.9.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
volcengine==1.0.193
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0
